fix: scanAndPopulate para que llame correctamente a populateFromFile doc: traducir todos los comentarios de inglés a español en CarDataPopulationService.java refactor: eliminar el Service redundante de GeneralCarPartsService
